# projet/shared/swagger.yaml
swagger: "2.0"
info:
  description: "This is a sample API"
  version: "V0.0.37"
  title: "CHOSA API"
host: "localhost:8080"
basePath: "/api"
...
schemes:
  - "http"
  - "https"
paths:
...
  /activities/:
    post:
      tags:
        - "activity-child"
      summary: "Create actuality"
      description: "This can only be done by the logged in user."
      operationId: "addActivity"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "Created activityChild object"
          required: true
          schema:
            $ref: "#/definitions/Activity"
      responses:
        default:
          description: "successful operation"
    get:
      tags:
        - "activity-child"
      summary: "Find All activities"
      description: "Returns all activities"
      operationId: "getAllActivities"
      produces:
        - "application/json"
      parameters:
        - name: offset
          description: offset liste
          type: "integer"
          in: query
          required: true
        - name: size
          description: The numbers of items to return
          type: string
          in: query
          required: true
      responses:
        200:
          description: "List of activitiesChild"
        404:
          description: "activitiesChild not found"
  /activities/{id}:
    patch:
      tags:
        - "activity-child"
      summary: "Update an existing activityChild"
      description: ""
      operationId: "activityUpdate"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: id
          description: activityChild id
          type: string
          in: path
          required: true
        - name: id
          description: activityChild properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/Activity"
      responses:
        200:
          description: "User updateted"
    get:
      tags:
        - "activity-child"
      summary: "Find activityChild by ID"
      description: "Returns a single activityChild"
      operationId: "getActivityById"
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "ID of activityChild to return"
          required: true
          type: "integer"
          format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Activity"
        404:
          description: "activityChild not found"
    delete:
      tags:
        - "activity-child"
      summary: "Delete activityChild"
      description: "This can only be done by the logged in user."
      operationId: "deleteActivity"
      produces:
        - "application/json"
      parameters:
        - name: "id"
          in: "path"
          description: "The name that needs to be deleted"
          required: true
          type: "string"
      responses:
        400:
          description: "Invalid id supplied"
        404:
          description: "activityChild not found"
...
securityDefinitions:
  Bearer:
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
...
  Activity:
    type: "object"
    required:
      - "age"
      - "status"
      - "date_publication"
    properties:
      id:
        type: "integer"
        format: "int64"
      age:
        type: "integer"
      status:
        type: "string"
        description: "Activity Status"
        enum:
          - "DRAFT"
          - "PUBLIC"
      date_publication:
        type: "string"
        format: "date"
      file:
        type: "integer"
        format: "int64"
    xml:
      name: "Activity"
...